+++ game/game.py
@@ -296,7 +296,7 @@ class Game:
         if self.turn > 1:
             self.conditions = self.generate_conditions()
 
-    def begin_turn_0(self) -> None:
+    def begin_turn_0(self, squadrons_start_full: bool) -> None:
         """Initialization for the first turn of the game."""
         from .sim import GameUpdateEvents
 
@@ -321,8 +321,9 @@ class Game:
                 # Rotate the whole TGO with the new heading
                 tgo.rotate(heading or tgo.heading)
 
-        self.blue.preinit_turn_0()
-        self.red.preinit_turn_0()
+        self.blue.preinit_turn_0(squadrons_start_full)
+        self.red.preinit_turn_0(squadrons_start_full)
+        # TODO: Check for overfull bases.
         # We don't need to actually stream events for turn zero because we haven't given
         # *any* state to the UI yet, so it will need to do a full draw once we do.
         self.initialize_turn(GameUpdateEvents())
@@ -381,7 +382,10 @@ class Game:
         self.red.bullseye = Bullseye(player_cp.position)
 
     def initialize_turn(
-        self, events: GameUpdateEvents, for_red: bool = True, for_blue: bool = True
+        self,
+        events: GameUpdateEvents,
+        for_red: bool = True,
+        for_blue: bool = True,
     ) -> None:
         """Performs turn initialization for the specified players.
 
@@ -433,9 +437,9 @@ class Game:
 
         # Plan Coalition specific turn
         if for_blue:
-            self.blue.initialize_turn()
+            self.blue.initialize_turn(self.turn == 0)
         if for_red:
-            self.red.initialize_turn()
+            self.red.initialize_turn(self.turn == 0)
 
         # Update cull zones
         with logged_duration("Computing culling positions"):